refactor: model/workflow.go를 entity/dto/runtime으로 파일 분리
281 LOC 단일 파일에 혼재되어 있던 세 가지 성격의 타입을 파일 단위로
분리. 패키지는 그대로 유지하므로 import path 변경은 없음.

- workflow_entity.go: GORM 영속화 대상
  (Workflow, WorkflowVersion, TaskDBModel, TaskGroupDBModel, TaskSnapshot
   + Workflow의 driver.Valuer / sql.Scanner)
- workflow_dto.go: API 요청/도메인 DTO
  (Task, TaskGroup, TaskDirectly, TaskGroupDirectly, Data,
   CreateXxxReq 계열, WorkflowSpecVersion 상수,
   Data/CreateDataReq의 driver.Valuer / sql.Scanner)
- workflow_runtime.go: Airflow 런타임 상태 (DB와 무관한 조회용)
  (WorkflowRun, TaskInstance, TaskInstanceReference, TaskLog,
   EventLog, EventLogs, TaskClearOption, WorkflowStatus)

이 분리로 Entity에 의도된 필드 변경이 API 응답을 건드릴 우려와 Airflow
런타임 타입이 DB 스키마로 오인될 여지를 구조적으로 분리.
